requirements-e2e.txt
pytest-playwright
playwright
Pillow
